Rename EventSchedule subcomponents
